requirements.txt
argh==0.26.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
blinker==1.4
certifi==2016.2.28
cffi==1.5.2
click==6.2
ConfigArgParse==0.10.0
construct==2.5.2
cryptography==1.1.2
enum34==1.1.2
futures==3.0.5
hpack==2.0.1
html2text==2015.11.4
idna==2.0
ipaddress==1.0.16
lxml==3.4.4
mitmproxy==0.15
ndg-httpsclient==0.4.0
netlib==0.15.1
passlib==1.6.5
pathtools==0.1.2
Pillow==3.0.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==0.15.1
pyparsing==2.0.7
pyperclip==1.5.26
PyYAML==3.11
requests==2.9.1
requests-futures==0.9.7
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
urwid==1.3.1
watchdog==0.8.3
wheel==0.24.0
